refactor(config): extract variant tests to sibling file
Moves 490 lines of tests from variant.rs to variant/tests.rs,
bringing variant.rs from 682 to 194 lines (well under the 500 soft limit).
